+++ src/app_optimizer/core/asMethodCalibrator.cpp
@@ -86,7 +86,7 @@ bool asMethodCalibrator::Manager()
     // Calibrate
     if (Calibrate(params)) {
         // Display processing time
-        wxLogMessage(_("The whole processing took %.3f min to execute"), float(sw.Time()) / 60000.0f);
+        wxLogMessage(_("The whole processing took %.3f min to execute"), static_cast<float>(sw.Time()) / 60000.0f);
 #if wxUSE_GUI
         wxLogStatus(_("Calibration over."));
 #endif
@@ -358,7 +358,7 @@ va1f asMethodCalibrator::GetClimatologyData(asParametersScoring *params)
 
     // Get start and end dates
     a1d predictandTime = m_predictandDB->GetTime();
-    auto predictandTimeDays = float(params->GetPredictandTimeHours() / 24.0);
+    auto predictandTimeDays = static_cast<float>(params->GetPredictandTimeHours() / 24.0);
     double timeStart, timeEnd;
     timeStart = wxMax(predictandTime[0], params->GetCalibrationStart());
     timeStart = floor(timeStart) + predictandTimeDays;
@@ -414,14 +414,14 @@ va1f asMethodCalibrator::GetClimatologyData(asParametersScoring *params)
     // Get index step
     double predictandTimeStep = predictandTime[1] - predictandTime[0];
     double targetTimeStep = params->GetTimeArrayTargetTimeStepHours() / 24.0;
-    int indexStep = int(targetTimeStep / predictandTimeStep);
+    int indexStep = static_cast<int>(targetTimeStep / predictandTimeStep);
 
     // Get vector length
     int dataLength = (indexPredictandTimeEnd - indexPredictandTimeStart) / indexStep + 1;
 
     // Process the climatology score
     va1f climatologyData(stationIds.size(), a1f(dataLength));
-    for (int iStat = 0; iStat < (int) stationIds.size(); iStat++) {
+    for (int iStat = 0; iStat < stationIds.size(); iStat++) {
         a1f predictandDataNorm = m_predictandDB->GetDataNormalizedStation(stationIds[iStat]);
 
         // Set data
@@ -753,7 +753,7 @@ bool asMethodCalibrator::SubProcessAnalogsNumber(asParametersCalibration &params
     vi analogsNbVect = params.GetAnalogsNumberVector(iStep);
 
     // Cannot be superior to previous analogs nb
-    int rowEnd = int(analogsNbVect.size() - 1);
+    int rowEnd = static_cast<int>(analogsNbVect.size() - 1);
     if (iStep > 0) {
         int prevAnalogsNb = params.GetAnalogsNumber(iStep - 1);
         if (prevAnalogsNb < analogsNbVect[analogsNbVect.size() - 1]) {
